+++ cpp_05/ex02/includes/Bureaucrat.hpp
@@ -6,7 +6,7 @@
 /*   By: Monsieur_Canard <Monsieur_Canard@studen    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/30 10:16:49 by Monsieur_Ca       #+#    #+#             */
-/*   Updated: 2024/07/30 15:37:03 by Monsieur_Ca      ###   ########.fr       */
+/*   Updated: 2024/08/02 11:33:45 by Monsieur_Ca      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,19 +28,19 @@ class AForm;
 class Bureaucrat {
 	private :
 		std::string const	_name;
-		unsigned int		_grade;
+		size_t				_grade;
 
 	public :
 		Bureaucrat();
-		Bureaucrat(std::string const name, unsigned int grade);
+		Bureaucrat(std::string const name, size_t grade);
 		Bureaucrat(Bureaucrat const &copy);
 		~Bureaucrat();
 
 		Bureaucrat &operator=(Bureaucrat const &rhs);
 
 		std::string const	getName() const;
-		unsigned int		getGrade() const;
-		void				setGrade(unsigned int grade);
+		size_t				getGrade() const;
+		void				setGrade(size_t grade);
 
 		void				incrementGrade();
 		void				decrementGrade();
